requirements.txt
pandas==1.2.4
matplotlib==3.3.4
streamlit==0.81.1
numpy==1.20.1
seaborn==0.11.1
Pillow==8.2.0
scikit_learn==0.24.2
networkx==2.5.1
numpy==1.21.1
pyvis==0.1.9